Move tests after rebase
