+++ registration/tests/test_forms.py
@@ -70,3 +70,58 @@ def test_step_three_accepts_valid_data():
         'email': 'john@jones.com',
     })
     assert form.is_valid()
+
+
+def test_company_profile_form_requires_name():
+    form = forms.CompanyBasicInfoForm(data={})
+
+    valid = form.is_valid()
+
+    assert valid is False
+    assert 'company_name' in form.errors
+    assert len(form.errors['company_name']) == 1
+    assert form.errors['company_name'][0] == 'This field is required.'
+
+
+def test_company_profile_form_requires_description():
+    form = forms.CompanyBasicInfoForm(data={})
+
+    valid = form.is_valid()
+
+    assert valid is False
+    assert 'description' in form.errors
+    assert len(form.errors['description']) == 1
+    assert form.errors['description'][0] == 'This field is required.'
+
+
+def test_company_profile_form_requires_website():
+    form = forms.CompanyBasicInfoForm(data={})
+
+    valid = form.is_valid()
+
+    assert valid is False
+    assert 'website' in form.errors
+    assert len(form.errors['website']) == 1
+    assert form.errors['website'][0] == 'This field is required.'
+
+
+def test_company_profile_form_rejects_invalid_website():
+    form = forms.CompanyBasicInfoForm(data={'website': 'google'})
+
+    valid = form.is_valid()
+
+    assert valid is False
+    assert 'website' in form.errors
+    assert len(form.errors['website']) == 1
+    assert form.errors['website'][0] == 'Enter a valid URL.'
+
+
+def test_company_profile_form_accepts_valid_data():
+    data = {'company_name': 'Amazon UK',
+            'website': 'http://amazon.co.uk',
+            'description': 'Ecommerce'}
+    form = forms.CompanyBasicInfoForm(data=data)
+
+    valid = form.is_valid()
+
+    assert valid is True
